+++ EasyDriver.Tests/Unit/XpathTest.cs
@@ -1,7 +1,9 @@
 ﻿using Comfast.EasyDriver.Se.Finder;
+using JetBrains.Annotations;
 
 namespace EasyDriver.Tests.Unit;
 
+[TestSubject(typeof(Xpath))]
 public class XpathTest {
     [Fact] public void EscapeTextTest() {
         DoTest("It's \"hard\" text 'to' match", "concat('It', \"'\", 's \"hard\" text ', \"'\", 'to', \"'\", ' match')");